poslednje izmene za FE

+++ src/main/java/com/iktpreobuka/services/PredmetDao.java
@@ -6,7 +6,7 @@ import com.iktpreobuka.entites.dto.PripadaRazredPredmetDTO;
 
 public interface PredmetDao {
 
-	public ResponseEntity<?> addRazred(Integer pred, PripadaRazredPredmetDTO razPred);
+	public ResponseEntity<?> addRazred(Long pred, PripadaRazredPredmetDTO razPred);
 
 	public ResponseEntity<?> addOcena(Integer ocena, PripadaRazredPredmetDTO ocenaPredmet);
 }
